add "torch.distributed.launch"
I was training yolov8 on multi-GPU devices on pytorch version 1.8. and got the following error. : /opt/conda/bin/python: No module named torch.distributed.run

Then I noticed that torch.distributed.run was introduced as a part of the development in PyTorch version 1.9. However, it was mentioned that we are free to use PyTorch version >1.7. Therefore, for those using PyTorch versions 1.7 and 1.8, the only option available to them is to use torch.distributed.run. However, in the current codebase, I did not find an option to use torch.distributed.launch. To address this, I have added a line of code which allows the use of torch.distributed.launch if the PyTorch version is less than 1.9.0.